requirements.txt
# Backend API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Data Processing
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Configuration
python-dotenv==1.1.0
PyYAML==6.0.2

# HTTP Client
httpx==0.25.2

# CORS
fastapi-cors==0.0.6